Added NavBar.js, refactored logout logic
