+++ src/components/Root.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { Router, Route } from 'react-router-dom';
 import Reset from '../styles/reset.css';
+import NavBar from './NavBar';
 import App from './App';
 import Home from './Home';
 import Callback from './Callback';
@@ -23,6 +24,7 @@ const Root = ({ store }) => {
         <React.Fragment>
           <Reset />
           <Route exact path="/" render={props => <App auth={auth} {...props} />} />
+          <Route path="/" render={props => <NavBar auth={auth} {...props} />} />
           <Route path="/home" render={props => <Home auth={auth} {...props} />} />
           <Route
             path="/callback"
